v1.101013 	Enh.		Operation expression. The first input is the expression, subsequent inputs are named. 	BUG:		The bitprecision was capped at 63, even though the error message says 64.
v1.101011
	Enh.		Operation random. takes 2 values representing the high/low values of the interval.
				so eg. precision 0 with inputs 1 and 2 will return either 1 or 2. 
				But precision 1 with the same inputs will return 1.0,1.1,1.2,1.3, ... 2.0
				The precision attribute affects the number of decimal points (and therefore the size of the interval)
	BUG:		In errors, mysql password was being shown. Now it is obscured with eg: p******d.

+++ core/iteration.cpp
@@ -72,7 +72,7 @@ operation(orig->operation),lastrow(orig->lastrow),expanded(orig->expanded),
 currentrow(orig->currentrow),numreps(orig->numreps),currentkey(orig->currentkey) {
 }
 unsigned long long Iteration::forcedbreak() const {
-	unsigned long long forced_break = INT_MAX;
+	unsigned long long forced_break = ULLONG_MAX;
 	string tmp_var;
 	if ( ItemStore::get("ITERATION_BREAK_COUNT",tmp_var) ) {
 		pair<unsigned long long,bool> forced_break_setting = String::znatural(tmp_var);
@@ -317,7 +317,7 @@ bool Iteration::operation_repeat() {
 	bool fully_evaluated = true;
 	string tmp_var;
 	string controlstring;
-	unsigned long long forced_break = INT_MAX;
+	unsigned long long forced_break = ULLONG_MAX;
 	if ( ! inputs.empty() ) { 
 		const DataItem* res = inputs[0]->results.result();
 		if (res != NULL) {
@@ -385,7 +385,7 @@ bool Iteration::operation_sql() {
 					if (definputs.size() > 0) {
 						DefInpType* base_template = definputs[0];
 						definputs.clear();
-						unsigned long long forced_break = INT_MAX;
+						unsigned long long forced_break = ULLONG_MAX;
 						long long queryrows = query->getnumrows();
 						numreps = queryrows < 1 ? 0 :  queryrows;
 						if ( ItemStore::get("ITERATION_BREAK_COUNT",tmp_var) ) {
